+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 - The `--last` flag to `copy` to control how many snapshots should be copied
 
+### Fixed
+
+- A minor issue in the `Libsecret` password prompt
+
 ## 20.0.0 - 2023-07-04
 
 ### Added