+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 ## Unreleased
 
+## 20.1.0 - 2023-08-29
+
 ### Added
 
 - The `--last` flag to `copy` to control how many snapshots should be copied